Add new type of quiz
do_i_know_it is a quiz designed around asking you the questions multiple
times, and removing it from the question pool only once you've gotten it
correct three times in a row.

THIS SCRIPT IS "DESTRUCTIVE" in that it will make changes to your JSON
file so you may want to create a copy.